Module rename: Internal.Linear => Linear.Internal
