+++ src/Control/Functor/Linear/Internal/MonadTrans.hs
@@ -3,11 +3,11 @@
 {-# LANGUAGE NoImplicitPrelude #-}
 {-# LANGUAGE QuantifiedConstraints #-}
 {-# LANGUAGE RankNTypes #-}
-module Control.Functor.Internal.Linear.MonadTrans
+module Control.Functor.Linear.Internal.MonadTrans
   ( MonadTrans(..)
   ) where
 
-import Control.Functor.Internal.Linear.Class
+import Control.Functor.Linear.Internal.Class
 
 class (forall m. Monad m => Monad (t m)) => MonadTrans t where
   lift :: Monad m => m a %1-> t m a
