+++ src/Data/Functor/Linear/Internal/Functor.hs
@@ -1,7 +1,7 @@
 {-# OPTIONS_HADDOCK hide #-}
 {-# LANGUAGE LinearTypes #-}
 {-# LANGUAGE NoImplicitPrelude #-}
-module Data.Functor.Internal.Linear.Functor
+module Data.Functor.Linear.Internal.Functor
   (
     Functor(..)
   , (<$>)
