+++ src/Data/Functor/Linear/Internal/Traversable.hs
@@ -9,7 +9,7 @@
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE TypeApplications #-}
 
-module Data.Functor.Internal.Linear.Traversable
+module Data.Functor.Linear.Internal.Traversable
   ( -- * Linear traversable hierarchy
     -- $ traversable
     Traversable(..)
@@ -17,11 +17,11 @@ module Data.Functor.Internal.Linear.Traversable
   , mapAccumL, mapAccumR
   ) where
 
-import qualified Control.Functor.Internal.Linear.Class as Control
-import qualified Control.Functor.Internal.Linear.State as Control
-import qualified Control.Functor.Internal.Linear.Instances as Control
-import qualified Data.Functor.Internal.Linear.Functor as Data
-import qualified Data.Functor.Internal.Linear.Applicative as Data
+import qualified Control.Functor.Linear.Internal.Class as Control
+import qualified Control.Functor.Linear.Internal.State as Control
+import qualified Control.Functor.Linear.Internal.Instances as Control
+import qualified Data.Functor.Linear.Internal.Functor as Data
+import qualified Data.Functor.Linear.Internal.Applicative as Data
 import Data.Functor.Const
 import Prelude.Linear.Internal
 import Prelude (Maybe(..), Either(..))
